removed prior from being added to logp
"Regarding chi^2, the "best fit" params and transit depths that Platon gives by default are not the highest likelihood but the highest posterior probability, i.e. they've been weighed by the prior. For MLE, this is not what you want. Go to combined_retriever.py and comment out where the prior is added to result.logp. This is only used for the best fit values so the posterior samples are not affected. "